requirements.txt
# Django и основные компоненты
Django==5.2.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==24.1
django-extensions==3.2.3
setuptools==69.2.0

# Документация API
drf-yasg==1.21.7

# База данных
mysqlclient==2.2.4
PyMySQL==1.1.0

# Работа с изображениями
Pillow==10.2.0

# Переменные окружения
python-dotenv==1.0.1

# Утилиты
Werkzeug==3.0.1

# Безопасность
django-environ==0.11.2

# Оптимизация
django-debug-toolbar==4.3.0

# Форматирование
black==24.1.1
isort==5.13.2

# Линтинг
flake8==7.0.0
pylint==3.0.3

# Тестирование
pytest==8.0.0
pytest-django==4.8.0 